+++ src/js/displayTask.js
@@ -1,17 +1,61 @@
-import { makeTaskDescriptionEditable, handleDeleteButtonClick } from './taskHandlers.js';
+let tasks = [];
 
-export let tasks = [];
 export const taskList = document.getElementById('todo-list');
 
 export const loadTasks = () => {
   const storedTasks = localStorage.getItem('tasks');
   if (storedTasks) {
-    tasks = JSON.parse(storedTasks);
+    tasks = JSON.parse(storedTasks); // Replace this line
   }
 };
 
 export const saveTasks = () => {
   localStorage.setItem('tasks', JSON.stringify(tasks));
+};
+
+export const makeTaskDescriptionEditable = (event) => {
+  const taskDescription = event.target;
+  taskDescription.setAttribute('contenteditable', 'true'); // Make the task description editable
+  taskDescription.focus(); // Focus on the task description element
+
+  const saveDescription = () => {
+    const newDescription = taskDescription.textContent.trim();
+    if (newDescription !== '') {
+      const listItem = taskDescription.parentElement;
+      const index = Array.from(listItem.parentNode.children).indexOf(listItem);
+      if (index !== -1) {
+        tasks[index].description = newDescription;
+        saveTasks();
+      }
+    }
+
+    taskDescription.setAttribute('contenteditable', 'false'); // Make the task description non-editable
+    taskDescription.removeEventListener('blur', saveDescription); // Remove the blur event listener
+  };
+
+  taskDescription.addEventListener('keydown', (event) => {
+    if (event.key === 'Enter') {
+      event.preventDefault(); // Prevent the default behavior of the Enter key
+      saveDescription();
+    }
+  });
+
+  taskDescription.addEventListener('blur', saveDescription);
+};
+
+export const handleDeleteButtonClick = (event) => {
+  const deleteButton = event.target;
+  const listItem = deleteButton.parentNode;
+  const index = Array.from(listItem.parentNode.children).indexOf(listItem);
+
+  tasks.splice(index, 1); // Remove the task from the array
+
+  // Adjust the index of remaining tasks
+  for (let i = index; i < tasks.length; i += 1) {
+    tasks[i].index = i + 1;
+  }
+
+  saveTasks();
 };
 
 export const populateTaskList = () => {
@@ -38,6 +82,7 @@ export const populateTaskList = () => {
     listItem.appendChild(taskDescription);
 
     const editButton = document.createElement('button');
+    const deleteButton = document.createElement('button');
     editButton.classList.add('edit-btn');
     editButton.addEventListener('click', () => {
       editButton.style.display = 'none'; // Hide the edit button
@@ -45,7 +90,6 @@ export const populateTaskList = () => {
     });
     listItem.appendChild(editButton);
 
-    const deleteButton = document.createElement('button');
     deleteButton.classList.add('btn-delete');
     deleteButton.addEventListener('click', () => {
       deleteButton.style.display = 'none'; // Hide the delete button
